almost there, one fixme to go
