+++ src/mpyl/steps/deploy/dagster.py
@@ -23,7 +23,9 @@ from .k8s.chart import ChartBuilder
 from .k8s.cluster import get_cluster_config_for_project
 from .k8s.helm import write_chart
 from .k8s.resources.dagster import to_user_code_values, to_grpc_server_entry, Constants
-from .. import Step, Meta, Input, Output
+from ..input import Input
+from ..output import Output
+from ..step import Step, Meta
 from ...utilities.dagster import DagsterConfig
 from ...utilities.docker import DockerConfig
 from ...utilities.helm import convert_to_helm_release_name, get_name_suffix
