+++ src/mpyl/steps/deploy/kubernetes_job.py
@@ -5,8 +5,9 @@ from logging import Logger
 from . import STAGE_NAME
 from .k8s import generate_helm_charts
 from .k8s.chart import ChartBuilder, to_cron_job_chart, to_job_chart
-from .. import Step, Meta
-from ..models import Input, Output
+from ..input import Input
+from ..output import Output
+from ..step import Meta, Step
 
 
 class DeployKubernetesJob(Step):
