+++ src/mpyl/steps/postdeploy/skip.py
@@ -5,9 +5,10 @@ Can be used to add the stage to the run plan, so you can run something else in y
 
 from logging import Logger
 
-from .. import Step, Meta
-from ..models import Input, Output
 from . import STAGE_NAME
+from ..input import Input
+from ..output import Output
+from ..step import Meta, Step
 
 
 class PostdeploySkip(Step):
