+++ src/mpyl/utilities/docker/__init__.py
@@ -3,10 +3,9 @@
 from dataclasses import dataclass
 from typing import Dict, Optional
 
-from ruamel.yaml import yaml_object, YAML
+from ruamel.yaml import YAML
 
 from ...project import Project
-from ...steps.models import Input
 
 yaml = YAML()
 
@@ -98,12 +97,6 @@ class DockerConfig:
             )
         except KeyError as exc:
             raise KeyError(f"Docker config could not be loaded from {config}") from exc
-
-
-def docker_image_tag(step_input: Input) -> str:
-    git = step_input.run_properties.versioning
-    tag = git.tag if git.tag else f"pr-{git.pr_number}"
-    return f"{step_input.project_execution.name.lower()}:{tag}".replace("/", "_")
 
 
 def registry_for_project(
